Cambios carrito historial

+++ Frontend/src/app/components/carrito/carrito.component.ts
@@ -2,10 +2,14 @@ import { CommonModule } from '@angular/common';
 import { Component } from '@angular/core';
 import { Router } from '@angular/router';
 import { UsuarioService } from '../../services/usuarios.service';
+import { FormBuilder, FormGroup, ReactiveFormsModule, Validators } from '@angular/forms';
+import Swal from 'sweetalert2';
+import { lastValueFrom } from 'rxjs';
+
 
 @Component({
   selector: 'app-carrito',
-  imports: [CommonModule],
+  imports: [CommonModule,ReactiveFormsModule],
   templateUrl: './carrito.component.html',
   styleUrl: './carrito.component.css'
 })
@@ -15,8 +19,18 @@ carrito: any[] = [];
   user: string = JSON.parse(localStorage.getItem('currentUser') || '{}').username;
   hayProductos: boolean = false;
   usuario: any;
-  constructor(private router: Router, private usuarioService: UsuarioService) {
-   
+
+  quiereComprar: boolean = false;
+  userForm: FormGroup;
+  submitted = false;
+  responseMessage = '';
+
+  constructor(private fb: FormBuilder,private router: Router, private usuarioService: UsuarioService) {
+   this.userForm = this.fb.group({
+      telefono: ['', [Validators.required,Validators.pattern(/^\+?[0-9]{4,15}$/)]],
+      departamento: ['', Validators.required],
+      direccion: ['', Validators.required]
+    });
     this.getCarrito();
     
   }
@@ -65,25 +79,48 @@ verProducto(id : string) {
 
 
   comprar() {
-    this.usuarioService.crearCompra(this.user,this.total).subscribe({
-      
-    });
+    this.quiereComprar = true;
+}
 
+get f() {
+    return this.userForm.controls;
+  }
+
+async onSubmit() {
+ 
+  this.submitted = true;
+  
+   if (this.userForm.invalid) return;
+
+    let telefonoNumber = this.userForm.get('telefono')?.value;
+    console.log(telefonoNumber);
+    try{
+    await lastValueFrom( this.usuarioService.crearCompra(this.user,this.total - 200,telefonoNumber,this.userForm.get('direccion')?.value,this.userForm.get('departamento')?.value));
     for (let i = 0; i < this.carrito.length; i++) {
       
-    this.usuarioService.comprarCarrito(this.user, this.carrito[i].id,this.carrito[i].precio,this.carrito[i].cantidad).subscribe({
-        
-});
+    await lastValueFrom (this.usuarioService.comprarCarrito(this.user, this.carrito[i].id,this.carrito[i].precio,this.carrito[i].cantidad));
 
   }
-  this.eliminarTodos();
-this.carrito.length = 0;
-
-
-      if(this.carrito.length === 0) {
-      this.hayProductos = false;
+  
+  Swal.fire({
+          icon: 'success',
+          title: '¡Compra realizada con éxito!',
+          showConfirmButton: false,
+          timer: 2000
+        }).then(() => {
+          
+          this.eliminarTodos();
+          this.carrito.length = 0;
+          this.userForm.reset();
+          this.submitted = false;
+          this.hayProductos = false;
+        });
+      }catch(error){
+        console.error("Error en la compra",error);
       }
-}
+
+       
+  }
 
 eliminarTodos(){
   for (let i = 0; i < this.carrito.length; i++) {
@@ -97,4 +134,9 @@ eliminarTodos(){
     });
   }
 }
+
+
+
+
+
 }
